Depth First Search
Depth first search is used to to traverse the graph. 
First it goes to one node then goes towards the depth till the dead end  and after that it backtracks.
It is not an algo to find the shortest path between source and destination.
It is used to find whether there exists a path between source to destination or not.